backend-example/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ccxt==4.1.64
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.4
pandas-ta==0.3.14b0
pyjwt==2.8.0
email-validator==2.1.0
python-multipart==0.0.6
